test range

+++ Steam_Hunters/Steam Hunters/Steam_Hunters/Entity/Enemies.cs
@@ -9,7 +9,7 @@ namespace Steam_Hunters
 {
     class Enemies : Entity
     {
-        public Enemies(Texture2D tex, Vector2 pos, Point frameSize, Point sheetSize, int Hp, int MaxHp, int Gold, int Item, int AttackRangeRadius, int SearchRadius, float MovementSpeed,
+        public Enemies(Texture2D tex, Vector2 pos, Point frameSize, Point sheetSize, int Hp, int MaxHp, int Gold, int Item, float AttackRangeRadius, float SearchRadius, float MovementSpeed,
          float AttackSpeed, float MapPos, bool Aggro, double AttackCooldown)
             : base(tex, pos, frameSize, sheetSize, Hp, MaxHp, Gold, Item, AttackRangeRadius, SearchRadius, MovementSpeed, AttackSpeed, MapPos, Aggro, AttackCooldown)
         {
@@ -19,9 +19,12 @@ namespace Steam_Hunters
         //    base.Update();
         //}
         public override void Draw(SpriteBatch spriteBatch)
-        {
+        {   
+           
             base.Draw(spriteBatch);
         }
+
+
     }
 }
 
@@ -29,7 +32,7 @@ namespace Steam_Hunters
 
 
 
-
+             
 //if (Aggro == true)
 //             {
 //                 if (target != null)
@@ -46,31 +49,29 @@ namespace Steam_Hunters
 
 //             }
 
-//  public bool IsInRange(Vector2 pos)
-//        {
-//            return Vector2.Distance(center, pos) <= radius;
-//        }
 
-//        public void GetClosestEnemy(List<Enemy> enemies)
-//        {
-//            target = null;
-//            float smallestRange = radius;
+        //public void GetClosestEnemy(List<Enemy> enemies)
+        //{
+        //    target = null;
+        //    float smallestRange = radius;
 
-//            foreach(Enemy e in enemies)
-//            {
-//                if (Vector2.Distance(center, e.Center) < smallestRange)
-//                {
-//                    smallestRange = Vector2.Distance(center, e.Center);
-//                    target = e;
-//                }
-//            }
-//        }
-//        protected void FaceTarget()
-//        {
-//            Vector2 direction = center - target.Center;
-//            direction.Normalize();
-//            rotation = (float)Math.Atan2(-direction.X, direction.Y);
-//        }
+        //    foreach(Enemy e in enemies)
+        //    {
+        //        if (Vector2.Distance(center, e.Center) < smallestRange)
+        //        {
+        //            smallestRange = Vector2.Distance(center, e.Center);
+        //            target = e;
+        //        }
+        //    }
+        //}
+        //protected void FaceTarget()
+        //{
+        //    Vector2 direction = center - target.Center;
+        //    direction.Normalize();
+        //    rotation = (float)Math.Atan2(-direction.X, direction.Y);
+        //}
+
+
 
 
 //if (EnemyPos >PlayerPos)
